backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langgraph>=0.0.69
langchain>=0.1.0
langchain-community>=0.0.20
groq>=0.4.0
pydantic>=2.0.0
httpx>=0.25.0
slowapi>=0.1.9
